% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinc.R
\name{BiomarkerData}
\alias{BiomarkerData}
\title{Biomarker Data}
\usage{
BiomarkerData(theData, biomarkerName, groupId, thresholds)
}
\arguments{
\item{theData}{the data}

\item{biomarkerName}{the biomarker}

\item{groupId}{the group}

\item{thresholds}{the thresholds}
}
\value{
output
}
\description{
Does some stuff
}
