% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Validation}
\usage{
validation(theData, biomarkerField, aggregationField, groupId, thresholds)
}
\arguments{
\item{theData}{the data}

\item{biomarkerField}{the field name of the biomarker measurement data
must be labelled "zinc", "haemoglobin", "ferritin", "stfr", "rbp", "retinol"
data will not be read.}

\item{aggregationField}{the field name to aggregate by}

\item{groupId}{the group}

\item{thresholds}{the thresholds}
}
\value{

}
\description{
Validates input files and parameters for the SummaryStats function
}
\details{
This function can be used to evaluate the input files ('theData' and
'thresholds') are complete and in the correct format.
It also checks the function paramaters (biomarkerField, aggregationField
and groupId) have been called correctly and match the input data.

If there is no error returned the parameters and datasets are correct, and
you may proceed with using the SummaryStats function.
}
