% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{SummaryStats}
\alias{SummaryStats}
\title{Biomarker Data}
\usage{
SummaryStats(
  theData,
  biomarkerField,
  aggregationField,
  groupId,
  thresholds,
  Flag_SurvWeightRun = FALSE,
  Flag_SurvWeightSupplied = FALSE,
  Flag_HaemAltAdjust = FALSE,
  Flag_SmokeAdjust = FALSE
)
}
\arguments{
\item{theData}{the data}

\item{biomarkerField}{the field name of the biomarker measurement data.
When a zero value is observed in the biomarkerField, there is a nominal value
of 0.001 added to the zero value in order to #' facilitate statistical
computation (in particular Log transform in the BRINDA package)}

\item{aggregationField}{the field name to aggregate by}

\item{groupId}{the group}

\item{thresholds}{the thresholds}

\item{Flag_SurvWeightRun}{If survey weights have already been run (?) set
Flag_SurvWeightRun = TRUE, else default is FALSE and the survey weights will
be run}

\item{Flag_SurvWeightSupplied}{If survey weights have already been supplied (?) set
Flag_SurvWeightupplied = TRUE, else default is FALSE}

\item{Flag_HaemAltAdjust}{If Haemogoblin adjustment for altitude has already
been applied set Flag_HaemAltAdjust = TRUE, else default is FALSE and the
Haemoglobin will be adjusted}

\item{Flag_SmokeAdjust}{If Haemogoblin adjustment for smoking status has already
been applied set Flag_HaemAltAdjust = TRUE, else default is FALSE and the
Haemoglobin will be adjusted}
}
\value{
output
}
\description{
Computes summary statistics for National Micronutrient Surveys
}
