% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{SummaryStats}
\alias{SummaryStats}
\title{Biomarker Data}
\usage{
SummaryStats(theData, biomarkerName, groupId, thresholds)
}
\arguments{
\item{theData}{the data}

\item{biomarkerName}{the biomarker}

\item{groupId}{the group}

\item{thresholds}{the thresholds}
}
\value{
output
}
\description{
Does some stuff
}
